move mop::util into mop::object and rename it to mop::object::util
This now makes mop::object a single self contained unit with no non-core dependencies that can be shipped independently if desired. 

This commit also makes mop::object backwards compatible to 5.6. Though it does this by using MRO::Compat so that does add a non-core dependency if you go back before 5.10 (which is when mro.pm was introduced).