Improving the MOP::Method constructor
As with MOP::Role/Class, there really is no slots associated with
this class, it simply blesses a ref to the thing it is introspecting
which in this case is the CODE ref. So in the spirit of this I
changed the constructor to me like the MOP::Role/Class version and
it now takes multiple argument forms.